requirements.txt
streamlit
azure-ai-vision-imageanalysis
azure-core
pillow
python-dotenv
requests